Disabled for windows

+++ community/src/test/java/org/neo4j/backup/TestBackupToolHa.java
@@ -26,6 +26,7 @@ import static org.neo4j.backup.TestBackupToolEmbedded.createSomeData;
 import static org.neo4j.backup.TestBackupToolEmbedded.runBackupToolFromOtherJvmToGetExitCode;
 import static org.neo4j.helpers.collection.MapUtil.stringMap;
 import static org.neo4j.kernel.Config.ENABLE_ONLINE_BACKUP;
+import static org.neo4j.kernel.Config.osIsWindows;
 import static org.neo4j.kernel.HighlyAvailableGraphDatabase.CONFIG_KEY_HA_MACHINE_ID;
 import static org.neo4j.kernel.HighlyAvailableGraphDatabase.CONFIG_KEY_HA_SERVER;
 import static org.neo4j.kernel.HighlyAvailableGraphDatabase.CONFIG_KEY_HA_ZOO_KEEPER_SERVERS;
@@ -53,6 +54,7 @@ public class TestBackupToolHa
     @Before
     public void before() throws Exception
     {
+        if ( osIsWindows() ) return;
         FileUtils.deleteDirectory( new File( PATH ) );
         FileUtils.deleteDirectory( new File( BACKUP_PATH ) );
 
@@ -77,6 +79,7 @@ public class TestBackupToolHa
     @After
     public void after() throws Exception
     {
+        if ( osIsWindows() ) return;
         for ( GraphDatabaseService instance : instances )
         {
             instance.shutdown();
@@ -87,6 +90,7 @@ public class TestBackupToolHa
     @Test
     public void makeSureBackupCanBePerformedFromCluster() throws Exception
     {
+        if ( osIsWindows() ) return;
         assertEquals( 0, runBackupToolFromOtherJvmToGetExitCode( "-full", "-from-ha", "localhost:2181", "-to", BACKUP_PATH ) );
         assertEquals( representation, DbRepresentation.of( BACKUP_PATH ) );
         DbRepresentation newRepresentation = createSomeData( instances.get( 2 ) );
